Fix walletconnect spamming websocket issue
